pyproject.toml
[project]
name = "comfyui-kokoro-remote"
description = "Remote-only Kokoro TTS nodes for ComfyUI (HTTP to your Kokoro server)."
version = "2.0.0"
license = { file = "LICENSE" }
dependencies = ["numpy", "requests", "python-dotenv"]

[project.urls]
Repository = "https://github.com/bigsk1/comfyui-kokoro-remote"

[tool.comfy]
PublisherId = "bigsk1"
DisplayName = "comfyui-kokoro-remote"
Icon = ""
